pyproject.toml
[project]
name = "nonebot-plugin-qq-maimaidx"
version = "0.0.7"
authors = [
  { name="Yuri-YuzuChaN", email="806235364@qq.com" },
]
description = "maimaidx plugin for nonebot2"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "nonebot2<3.0.0,>=2.0.0",
    "nonebot-adapter-qq<2.0.0,>=1.3.4",
    "nonebot-plugin-apscheduler<1.0.0,>=0.3.0",
    "aiofiles<24.0.0,>=23.2.1",
    "httpx<1.0.0,>=0.23.1",
    "pillow<11.0.0,>=10.0.0",
    "pydantic<2.0.0,>=1.10.13",
    "pyecharts<3.0.0,>=2.0.0",
    "snapshot_phantomjs<1.0.0,>=0.0.1",
    "numpy<2.0.0,>=1.26.2"
]

[project.urls]
"Homepage" = "https://github.com/Yuri-YuzuChaN/nonebot-plugin-qq-maimaidx"
"Bug Tracker" = "https://github.com/Yuri-YuzuChaN/nonebot-plugin-qq-maimaidx/issues"

[tool.setuptools.packages.find]
where = ["nonebot_plugin_qq_maimaidx"]

[tool.setuptools.package-data]
"*" = ["maimaidxhelp.png"]